Add --use-webgpu-adapter arg to WebGPU CTS harness
The arg is forwarded to the browser on startup which
will change the adapter returned by WebGPU to
requestAdapter. Passing the flag to the harness
is reflected in the platform tags so that suppressions
for a particular WebGPU adapter may be added.

Also rename --is-backend-validation to
--enable-dawn-backend-validation to match the Chrome
flag.

Bug: 1315703
Change-Id: I2b22c42f66d3c3560de52ca3cf8b9ebdd187598c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3580636
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Austin Eng <enga@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991849}