Populate the WebAuthnRequestId for Lacros windows
To support Lacros WebAuthn, we have to be able to retrieve the
WebAuthnRequestId from Lacros windows in ash. Therefore, we choose to
use the app_id of Lacros windows. Populate the WebAuthnRequestId as the
app_id for Lacros windows in ExoAppTypeResolver. Lacros UV WebAuthn
hasn't been enabled in this CL.

Bug: 1269528
Test: CQ
Test: u2fd tast tests

Change-Id: I884f7bb25903b767a474b45eef0ebc527aa47d27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3558697
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Cr-Commit-Position: refs/heads/main@{#991850}