[Intents] Implement In-Product Help Bubble for Intent Chip
The Intent Chip feature allows users to capture and open webpages in
relevant apps that they already have installed on their devices. This CL
adds a browser user education feature (in-product help bubble) that
brings attention to the Intent Chip and explains the chip's
functionality in an attempt increase awareness and usage of the chip
where possible.

Bug: 1304529
Change-Id: Ie7b6b57e71a93a171851d64736ffa958733918c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3564824
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Vivian Pao <vpao@google.com>
Cr-Commit-Position: refs/heads/main@{#991910}