arc-a11y: Port child node sorting logic from TalkBack
This CL implements a sorting of child nodes with a similar logic to
TalkBack. This makes navigation order in ChromeVox almost the same as
TalkBack.
One large difference after this CL is the use of traversalBefore/After
attribute, which is tracked in b/219840172.

Note that the logic has a small drawback. When there is a scrollable
content that can overlap with another content after scrolling, there is
a chance that scrolling makes the sorting result different. But this is
the same behavior as TalkBack, so it's better than having different
traversal orders between ChromeVox and TalkBack.

AX-Relnotes: ChromeVox behaves more like TalkBack when navigating within Android apps.
Bug: b/183571695
Test: AXTreeSourceArcTest
Change-Id: I0aa8b77e3ddf9252f4c51c044b07847c98ec3c5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577863
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991911}