[Fenced Frames] replacement of the initial empty document
There are different behaviors based on the InitialNavigationEntry
flag, which controls whether we create the initial NavigationEntry
on FrameTree creation or not.

Without InitialNavigationEntry, the initial empty document's entry
always gets replaced on subframes. Meanwhile, for main frames, it's
only replaced on the first navigation from the initial empty document.
So if the main frame stays on the initial empty document after the
first navigation (i.e. the navigation is the same document), the
initial empty document's entry won't get replaced on the second
navigation. And there's some difference between window.open() vs
window.open(about:blank).

With InitialNavigationEntry, the initial empty document's entry
always gets replaced on main frames and subframes. However, there's
a special case for main frames, related to session restore
(when you restart Chrome etc). Technically when you start a new tab,
you are showing the initial empty document, and our browser-side
tracking of the "is on initial empty document" state also reflects
that. However, a restored tab might already have restored
NavigationEntries for non-initial-empty-documents even though its
"is on initial empty document" state is true. In this case, we
shouldn't do the replacement on the next navigation, as otherwise,
we'll replace a restored non-initial-empty-document NavigationEntry
on the next navigation (see crbug.com/1295723#c42). So, for main
frame navigations, we need to replace iff the NavigationEntry is for
the initial empty document, i.e. it is an "initial NavigationEntry",
which is created on FrameTree creation time.  Meanwhile, for
subframe navigations, we can just check the "is on initial empty
document" state, as they don't have the session restore problem
(and initial empty document for subframes are not always on the
"initial NavigationEntry"). For fenced frames, we also don't have
the session restore problem, so we can just check the "is on
initial empty document" state that we set as true on FrameTree
creation time. So, when the fenced frame is created, it starts on the
initial empty document and it will be replaced when the navigation is
committed by setting src as “about:blank”.

Even if we have the "is on initial empty document" state as true,
we also need to explicitly treat fenced frames as subframes. So,
this CL changes IsMainFrame to GetParentOrOuterDocument in
ShouldReplaceCurrentEntryForNavigationFromInitialEmptyDocumentOrEntry().

BUG=1215096, 1190019

Change-Id: I4efa1748b3d13393d803b9ee5ad48049291e37c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3529286
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Cr-Commit-Position: refs/heads/main@{#991925}