[WebChannels] Fix crash when swipe to unfollow
Currently we see crashes when user want to unfollow a web channel
from the follow management page. It's because the
FollowedWebChannelItem doesn't retains the FollowedWebChannel, so
when looking for the corresponding webChannel from the item, it
returns nil. This cl changes the property attribute to strong,
so the item will retain the WebChannel object.

Bug: 1264872
Change-Id: I2adb198a7019f9beb8f01d48e377f4a7ba63ff4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3584782
Reviewed-by: edchin <edchin@google.com>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991940}