[Lacros] SigninManager::StartWebSigninFlow() completion callback
This CL adds a callback parameter to
SigninManager::StartWebSigninFlow(). This callback is invoked after
WebSigninHelperLacros completes adding a new account, either
successfully or not. The callback has a CoreAccountId parameter that
is empty if the flow fails.

This CL also adds one caller that benefits from the callback.
ShowExtensionSigninPrompt() might enable sync at the end of the account
addition flow if its `enable_sync` parameter is true.

Bug: 1260291
Change-Id: Ia6301112a29e7f86b2f4b0902796e5a4a62472c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582127
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991990}