[signin][Lacros] Support signin from extension context
This CL updates the CHROME_ID_CONSISTENCY_STATE cookie in the extension
context, to prevent the signin page from getting stuck on "Please Wait".

The value of the cookie for extension is simply mirrored from the value
in the main browsing context, as managing multiple versions of the
cookie seems overkill.

Bug: 1311042
Change-Id: I7ac564a91a2e192ea7cca41dc06b1892070b71ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3569069
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991995}