[Printing] Have PrintJob take in PrintJobManager
The removal of the PRINT_JOB_EVENT notification is blocked by listeners
that listen for the notification from *all* PrintJobs. This CL has
PrintJob take in the global PrintJobManager instance. A followup CL
will then have it directly call into PrintJobManager where it currently
fires the notification, allowing PrintJobManager to have knowledge of
all PrintJob instances without requiring on the notification (which it
currently does). This will in turn allow other classes to observe
events about all PrintJobs via PrintJobManager.

The one wrinkle is TestPrintJob, which is somewhat widely-used. This
CL allows PrintJob to be constructed with a null PrintJobManager
instance in that testing context only. As we transition PrintJobManager
to observing PrintJobs via the direct callback rather than the
notification, we will handle any tests that require that TestPrintJob
actually notify PrintJobManager of its creation (it's not clear a
priori whether there are any such tests).

Bug: 796051
Change-Id: I5b3d005a2ed0c2be2fda7941e7d7047f0334dc0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579111
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992007}