arc-a11y: Compute text only from nodes with importance
This change enforces recursive accessibility name computation in ARC
tree uses only text from nodes with accessibility importance.

For example, consider a following tree:
+ a parent node which can be focused, but without any text
++ a child node with text

If both nodes have importance properties, then we compute the name of
the parent node from the child one. When the parent is important but
the child is not, we don't want to use text in the child. Before this
CL, the text in the child is used.

This CL adds an explicit check to prevent the above.

AX-Relnotes: n/a

Fixed: b/227534934
Test: AccessibilityNodeInfoDataWrapperTest
Test: manually with TalkBack test app.
Change-Id: I75c2d0194549ad43ff0af2e4376c75d845616644
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3578509
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992013}