[Android][UPM] Create histograms for PasswordStoreBuiltInBackend
The Latency and Success metrics for the methods of the built-in
backend have been reported, but the histogram was not updated
accordingly.

The histogram specifications for Latency and Success metrics that
used to cover only the methods of Android backend are now
changed to be more generic. They now cover the cases of storing
the metrics for the built-in backend and Android backend, both
together on one histogram and separately.

Here is the list of histograms added by this CL:
http://go/paste/5516347460026368
It's also visible that 0 histograms were removed or made obsolete.

Overall metrics are introduced to facilitate the analysis of A/B
experiment results  and compare the performance of the backend
for both sets of users.

Backend-specific metrics are introduced because each group of
users will use both backends during MVP and we need to know the
performance for each of the backends.

Bug: 1298467
Change-Id: I931f59f9b213fc6a6a5cc8cdfb3aea539832e4c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3581031
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#992037}