[signin][Lacros] Fix tests to support non-syncing profiles
In preparation to enabling the LacrosNonSyncingProfiles feature, this CL
fixes a few unittest to support that mode.

OneGoogleBarLoaderImplTest is updated to support the "Mirror Landing"
flavor of the Mirror signin protocol, which uses
"consistency_enabled_by_default=true".

IdentityManagerTest is updated to assume kSignin as a default consent
level rather than kSync on Lacros.

Bug: 1260291
Change-Id: I55b3881b68a0066bebe8f9e7c7fbbd1ba54a2e93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582580
Reviewed-by: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: Monica Basta <msalama@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992084}