[DevTools] Don't interrupt V8 execution for Runtime.getProperties.
V8 (or the V8 inspector) provide no guarantees that harvesting
properties from the inspector while paused on an interrupt is
going to work or not. Most of the times this will likely work,
but the crashes indicate that there are cases where it doesn't.

Fixed: chromium:1277170
Change-Id: I9eaad5a736f9b99789f04a59b245bea2544ba620
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582582
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992103}