[iOS] Inset new popup to the omnibox
This CL updates the popup in Treatment 2 to inset itself to the edges of
the omnibox. It does this by having the `OmniboxPopupHostingController`
(in UIKit) look at the omnibox layout guide and pass the correct sizes
to the SwiftUI PopupView via the model.

It also leaves Treament 1 with no inset.

I considered making another "model" class to hold view data, as putting
the two space values in PopupModel feels a bit weird, but in the end,
I decided that wasn't necessary.

Change-Id: I17f6cd4e1ec79b9fea188e52bdf76a1ded10ad7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3584420
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#992106}