personalization: Use existing code paths to ensure proper caching and default wallpaper behavior.
Besides the need for a callback, `OnGooglePhotosWallpaperDecoded` and
`OnWallpaperDecoded` have identical logic. Since the former is used in
multiple places, the cleanest solution was just to run the callback and
call `OnWallpaperDecoded`.

Also added unit tests to ensure caching and default wallpaper are
happening.

Bug: b:228884203
Change-Id: I44acd2767b7b46d133d55a90b0ea3721d12d0b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3584285
Reviewed-by: David Black <dmblack@google.com>
Reviewed-by: Colin Kincaid <ckincaid@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Angus McLean <angusmclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992114}