NavigationControllerImpl::GetNavigationApiHistoryEntryVectors() uses the wrong index for initial history navs
When committing a history navigation in a new child frame, we are doing
a "replace" navigation, not appending to the session history.
NavigationControllerImpl::GetNavigationApiHistoryEntryVectors()
wrongly treats this case as an append and sends incorrect
NavigationApiHistoryEntryArrays to the renderer.

In the case of a same-document back/forward navigation, the Navigation
API verifies that its entry arrays contain the navigation destination
(in order to ensure that its own bookkeeping remains correct), and
cancels the navigation if it can't be found (believing that we have
hit a race condition and are attempting to navigate to a history
entry that is no longer in the back/forward list). Because the
Navigation API checks all navigations, not just ones that were started
via the Navigation API, sending incorrect
NavigationApiHistoryEntryArrays can lead to future same-document back/forward navigations getting spuriously cancelled.

Without this CL and with the NavigationApi enabled by default, the
following tests fail:
NavigationControllerBrowserTest.BackSameDocumentAfterBlockedSubframe
WebBundleNetworkBrowserTest.IframeSameDocumentNavigation

Bug: 1183545
Change-Id: I2d5b75c4690589dfe4673dc80aaeb0ac99e4b784
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3578222
Auto-Submit: Nate Chapin <japhet@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992118}