Fenced Frames: Renderer side event reporting
This CL plumbs the reporting metadata down to the renderer on commit and
implements the renderer process side changes for the event reporting
design. When `window.fence.reportEvent()` is called, a beacon is sent
from the renderer process reusing the existing implementation for
`navigatior.sendBeacon()`.

`window.fenced.reportEvent()` is allowed to be called inside fenced
frames and nested iframes inside the fenced frame tree.

https://github.com/WICG/turtledove/blob/main/Fenced_Frames_Ads_Reporting.md

Change-Id: I222745c452eb7cddd8b79ef19b90686ee2d48d8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3565244
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992145}