avoid unnecessary IPC on mac when painting
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/2923689
with fixes.

RootCompositorFrameSinkImpl::DisplayDidReceiveCALayerParams is called
after every frame. This in turn calls to OnDisplayReceivedCALayerParams.
If the CALayerParams haven't changed, then this IPC should not be
necessary. This IPC also has a side effect of updating the vsync, which
can potentially change from frame to frame. NativeWidgetMacNSWindowHost
queries DisplayLinkMac for vsync, but DisplayLinkMac only queries
every 10 seconds. So, this patch sends the IPC if the params haven't
changed, but 10 seconds has elapsed.

DisplayLinkMac queries the system for vsync on a background thread, so
that the first time through the vsync isn't available. To handle this
case I added support for a callback once vsync is enabled. This
ensures the vsync is properly updated the first time through.

The changes roughly result in something like the following:

Browser IO thread:
 . Before: https://i.imgur.com/Xa5tcIb.png
 . After: https://i.imgur.com/VM6wQTJ.png
Browser main thread:
 . Before: https://i.imgur.com/w4boRQJ.png
 . After: https://i.imgur.com/zJHdzEp.png

Change-Id: Ida80358f8331a440615d7c1944cd9d7f44e7cbe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3580129
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992183}