Don't hold async draw callbacks often.
Chrome tries to use async draw callbacks to signal when it's okay to
start an animated transition in android, for example when the surface
size changes.  Unfortunately, this can cause the android framework
to hang the UI thread during ViewRootImpl traversals.

Luckily, for some value of "luckily", the async draw callbacks don't
actually prevent the incorrectly sized frames from showing up on the
screen; they just block the window manager transaction.  So, except
for when we're showing the initial surface, there will still be a
mismatch between a resized surface and the GraphicBuffer.

That's why there are often one frame glitches during transitions;
the async draw callback signals the end of the WM animation / scale
that's trying to match the old-sized buffers to the new-sized
display surface.

So, if chrome posts a new-sized buffer before calling the async
draw callback, the WM scaling will be incorrectly applied.  If,
however, it calls the async draw callback before posting a new-sized
buffer, then the WM scale will be incorrectly removed from the
most recent old-sized buffer.  Either way, it triggers a draw and
it's unsynchronized.

So, we now are much less likely to hold async draw callbacks if
we think that the window is already on-screen; it doesn't help to
make it less janky, but is much less likely to hang the browser.

Bug: 1267336, 1270209
Change-Id: Iff4f6db32b48b643ae2bc13dc101921087dde9cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3543312
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992249}