[Metric] A general class to record feature discovery duration
This CL implements a general class to record feature discovery duration.
Please read the design doc attached to the issue for further details.

This CL does the following things:

(1) Create a class called `FeatureDiscoveryDurationReporter`. It is
a singleton and can be accessed from both Ash and browser.

(2) Users of this new class call `MaybeActivateObservation()` to
start an observation and call `MaybeFinishObservation()` to end an
observation. `MaybeActivateObservation()` and `MaybeFinishObservation()`
do not have to be paired. `FeatureDiscoveryDurationReporter` guarantees
that a feature's discovery time is recorded only once for the primary
user on one device.

(3) `FeatureDiscoveryDurationReporterImpl` is active only for primary
accounts. Ideally, feature discovery duration should be collected from
the users with little knowledge of new features. It is assumed to be
more likely for users to learn new features on primary accounts than
secondary accounts. If this assumption is held, the discovery duration
data collected from secondary accounts is distorted by users' prior
knowledge of new features gained from primary accounts. Therefore, data
collection on secondary accounts is banned.

(4) `FeatureDiscoveryDurationReporter` guarantees that the session
inactive duration (for example, the time a user spends when the screen
is locked) is not counted into the total discovery duration.

(5) `FeatureDiscoveryDurationReporter` takes multi-profile into
consideration. It guarantees that when the active user changes, the
inactive users feature discovery time counting is halted.

Bug: 1311344
Change-Id: I0ecbfcdaf710ecb4e54a8fa977482f2f67a3d940
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3551563
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992266}