Refactor DownloadBubbleRowView to decouple progress bar and cancel button
In deep scanning state, the progress bar will be associated with the
"Open now" button instead of the "Cancel" button, so we need to decouple
the progress bar and the cancel button.

Also refactor the way buttons are structured. Instead of creating and
deleting buttons on state changes, simply set the visibility of the
button. This pattern is also used in the old DownloadItemView.

No functional change.

Bug: 1311867
Change-Id: I4be64b289d8f7ad4e9303e635b3eb1c81169699f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577916
Reviewed-by: Rohit Bhatia <bhatiarohit@google.com>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992267}