Remove unused kSRTPromptGroupNameParam
The contents of this param are added to the software_reporter_tool.exe
command line, but the code to read that command-line flag is long gone.
It was used when experimenting with different prompt behaviours, to
record the experiment group name in logs collected from the reporter.

Bug: 1305047
Change-Id: Ida043b88834917652d3087f006d92fba3d71fa78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3585029
Auto-Submit: Joe Mason <joenotcharles@google.com>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992314}