Change crash_reporter:ChildExitObserver back to non-singleton.
In the distant past, ChildExitObserver had to be singleton as it was
managed by breakpad and accessed from various locations across
the code, so a singleton made sense.

Now, it's only accessed from one place in browser_main_parts so
can be converted back to a unique_ptr whose lifetime can be managed
more tightly.

This change allows a DCHECK to finally be restored in
BrowserChildProcessHostImpl::RemoveObserver as the entire lifetime
of this object is now within the lifetime of the task environment
of RenderViewHostTestHarness meaning OutOfMemoryReporterTest passes.

BUG=167126,1315700

Change-Id: Ied071d64e482411d38b9abfbd677165823af6855
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3349784
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Junbo Ke <juke@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992319}