Restore accessibility busy state exposure on documents
crrev.com/c/3268209 introduced a regression for NVDA users:
When a page is newly-loaded, NVDA no longer announced the title
or did an automatic SayAll.

This commit reverses the part of the change that introduced the
regression. The reason this change was originally made is due to the
fact that crrev.com/c/3268209 added the kBusy state to all documents
that were in the process of loading; not just the initial blank document
on Windows. However, exposing STATE_SYSTEM_BUSY on documents that are
still loading can prevent NVDA from creating its virtual buffer during
page load.

This commit does NOT reverse the addition of the kBusy state on Windows.
As a result, there is still the risk that NVDA will not create the
virtual buffer until the page load has completed.

AX-Relnotes: Fixed regression causing newly-loaded pages from being
presented by NVDA.

Bug: 1315210
Change-Id: Iaebfdf7a8c8c3a7b6ed1e689daf219b86a6121ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3585327
Auto-Submit: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#992355}