[BatteryLevelProvider] Use weak ptrs in Windows implementation
A major difference between the Windows and the Mac implementation of
BatteryLevelProvider is that the former is async. Thus, callers have to
always pass a weak pointer if they can't guarantee they will still be
alive when the callback is invoked.

With this change, it is always safe to use BatteryLevelProvider with an
unretained pointer.

Change-Id: Ie1fde8de1377ca47b19ce385442836842ff26790
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579511
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992370}